+++ benchexec/cgroups.py
@@ -215,7 +215,19 @@ def _parse_proc_pid_cgroup(content):
             yield (subsystem, path)
 
 
-def kill_all_tasks_in_cgroup(cgroup, ensure_empty=True):
+def _force_open_read(filename):
+    """
+    Open a file for reading even if we have no read permission,
+    as long as we can grant it to us.
+    """
+    try:
+        return open(filename, "rt")
+    except OSError:
+        os.chmod(filename, stat.S_IRUSR)
+        return open(filename, "rt")
+
+
+def kill_all_tasks_in_cgroup(cgroup):
     tasksFile = os.path.join(cgroup, "tasks")
 
     i = 0
@@ -225,7 +237,7 @@ def kill_all_tasks_in_cgroup(cgroup, ensure_empty=True):
         # SIGKILL. We added this loop when killing sub-processes was not reliable
         # and we did not know why, but now it is reliable.
         for sig in [signal.SIGKILL, signal.SIGINT, signal.SIGTERM]:
-            with open(tasksFile, "rt") as tasks:
+            with _force_open_read(tasksFile) as tasks:
                 task = None
                 for task in tasks:
                     task = task.strip()
@@ -240,7 +252,7 @@ def kill_all_tasks_in_cgroup(cgroup, ensure_empty=True):
                         )
                     util.kill_process(int(task), sig)
 
-                if task is None or not ensure_empty:
+                if task is None:
                     return  # No process was hanging, exit
             # wait for the process to exit, this might take some time
             time.sleep(i * 0.5)
@@ -463,17 +475,40 @@ class Cgroup(object):
         Kill all tasks in this cgroup and all its children cgroups forcefully.
         Additionally, the children cgroups will be deleted.
         """
+        # In this method we should attempt to guard against child cgroups
+        # that have been created and manipulated by processes in the run.
+        # For example, they could have removed permissions from files and directories.
 
-        def kill_all_tasks_in_cgroup_recursively(cgroup, delete):
-            for dirpath, dirs, _files in os.walk(cgroup, topdown=False):
-                for subCgroup in dirs:
-                    subCgroup = os.path.join(dirpath, subCgroup)
-                    kill_all_tasks_in_cgroup(subCgroup, ensure_empty=delete)
+        def recursive_child_cgroups(cgroup):
+            def raise_error(e):
+                raise e
 
-                    if delete:
-                        remove_cgroup(subCgroup)
+            try:
+                for dirpath, dirs, _files in os.walk(
+                    cgroup, topdown=False, onerror=raise_error
+                ):
+                    for subCgroup in dirs:
+                        yield os.path.join(dirpath, subCgroup)
+            except OSError as e:
+                # some process might have made a child cgroup inaccessible
+                os.chmod(e.filename, stat.S_IRUSR | stat.S_IXUSR)
+                # restart, which might yield already yielded cgroups again,
+                # but this is ok for the callers of recursive_child_cgroups()
+                yield from recursive_child_cgroups(cgroup)
 
-            kill_all_tasks_in_cgroup(cgroup, ensure_empty=delete)
+        def try_unfreeze(cgroup):
+            freezer_file = os.path.join(cgroup, "freezer.state")
+            try:
+                util.write_file("THAWED", freezer_file)
+            except OSError:
+                # Somebody could have fiddle with permissions, try to set them.
+                # If we are not owner, this also fails, but then there is nothing we
+                # can do. But the processes inside the run cannot change the owner.
+                try:
+                    os.chmod(freezer_file, stat.S_IRUSR | stat.S_IWUSR)
+                    util.write_file("THAWED", freezer_file)
+                except OSError:
+                    pass
 
         # First, we go through all cgroups recursively while they are frozen and kill
         # all processes. This helps against fork bombs and prevents processes from
@@ -484,16 +519,29 @@ class Cgroup(object):
         if FREEZER in self.per_subsystem:
             cgroup = self.per_subsystem[FREEZER]
             freezer_file = os.path.join(cgroup, "freezer.state")
-
             util.write_file("FROZEN", freezer_file)
-            kill_all_tasks_in_cgroup_recursively(cgroup, delete=False)
+
+            for child_cgroup in recursive_child_cgroups(cgroup):
+                with _force_open_read(os.path.join(child_cgroup, "tasks")) as tasks:
+                    for task in tasks:
+                        util.kill_process(int(task))
+
+                # This cgroup could be frozen, which would prevent processes from being
+                # killed and would lead to an endless loop below. cf.
+                # https://github.com/sosy-lab/benchexec/issues/840
+                try_unfreeze(child_cgroup)
+
             util.write_file("THAWED", freezer_file)
 
         # Second, we go through all cgroups again, kill what is left,
         # check for emptiness, and remove subgroups.
         # Furthermore, we do this for all hierarchies, not only the one with freezer.
         for cgroup in self.paths:
-            kill_all_tasks_in_cgroup_recursively(cgroup, delete=True)
+            for child_cgroup in recursive_child_cgroups(cgroup):
+                kill_all_tasks_in_cgroup(child_cgroup)
+                remove_cgroup(child_cgroup)
+
+            kill_all_tasks_in_cgroup(cgroup)
 
     def has_value(self, subsystem, option):
         """