+++ benchexec/tools/pichecker.py
@@ -10,11 +10,15 @@ import benchexec.tools.cpachecker as cpachecker
 
 class Tool(cpachecker.Tool):
     """
-    Tool info for Graves-CPA.
-    URL: https://github.com/will-leeson/cpachecker
+    Tool info for PIChecker.
+    URL: https://gitlab.com/Lapulatos/pichecker
     """
 
     REQUIRED_PATHS = list(cpachecker.Tool.REQUIRED_PATHS) + ["resources"]
 
+    def version(self, executable):
+        version = self._version_from_tool(executable, "-help", line_prefix="PIChecker")
+        return version.split("(")[0].strip()
+
     def name(self):
-        return "Graves-CPA"
+        return "PIChecker"